refactor: organize page objects by concern
