+++ pageObjectModels/dashboards/dashboardPage.ts
@@ -1,5 +1,5 @@
 import { Page } from '@playwright/test';
-import { BasePage } from './basePage';
+import { BasePage } from '../basePage';
 
 export class DashboardPage extends BasePage {
   readonly path: string;
