+++ pageObjectModels/users/editUserAdminPage.ts
@@ -1,5 +1,5 @@
 import { Page } from '@playwright/test';
-import { BASE_URL } from '../playwright.config';
+import { BASE_URL } from '../../playwright.config';
 import { UserAdminPage } from './userAdminPage';
 
 export class EditUserAdminPage extends UserAdminPage {
