+++ pageObjectModels/users/userAdminPage.ts
@@ -1,6 +1,6 @@
 import { Locator, Page } from '@playwright/test';
-import { ReferenceFieldGrid } from '../componentObjectModels/controls/referenceFieldGrid';
-import { BaseAdminFormPage } from './baseAdminFormPage';
+import { ReferenceFieldGrid } from '../../componentObjectModels/controls/referenceFieldGrid';
+import { BaseAdminFormPage } from '../baseAdminFormPage';
 
 export class UserAdminPage extends BaseAdminFormPage {
   readonly generalTabButton: Locator;
