+++ tests/app/app.spec.ts
@@ -4,18 +4,18 @@ import { UserFactory } from '../../factories/userFactory';
 import { test as base, expect } from '../../fixtures';
 import { UserStatus } from '../../models/user';
 import { AdminHomePage } from '../../pageObjectModels/adminHomePage';
-import { AppAdminPage } from '../../pageObjectModels/appAdminPage';
+import { AppAdminPage } from '../../pageObjectModels/apps/appAdminPage';
+import { AppsAdminPage } from '../../pageObjectModels/apps/appsAdminPage';
+import { AddGroupAdminPage } from '../../pageObjectModels/groups/addGroupAdminPage';
+import { EditGroupAdminPage } from '../../pageObjectModels/groups/editGroupAdminPage';
+import { GroupsSecurityAdminPage } from '../../pageObjectModels/groups/groupsSecurityAdminPage';
+import { AddRoleAdminPage } from '../../pageObjectModels/roles/addRoleAdminPage';
+import { EditRoleAdminPage } from '../../pageObjectModels/roles/editRoleAdminPage';
+import { RolesSecurityAdminPage } from '../../pageObjectModels/roles/rolesSecurityAdminPage';
+import { AddUserAdminPage } from '../../pageObjectModels/users/addUserAdminPage';
+import { EditUserAdminPage } from '../../pageObjectModels/users/editUserAdminPage';
+import { UsersSecurityAdminPage } from '../../pageObjectModels/users/usersSecurityAdminPage';
 import { AnnotationType } from '../annotations';
-import { AddGroupAdminPage } from './../../pageObjectModels/addGroupAdminPage';
-import { AddRoleAdminPage } from './../../pageObjectModels/addRoleAdminPage';
-import { AddUserAdminPage } from './../../pageObjectModels/addUserAdminPage';
-import { AppsAdminPage } from './../../pageObjectModels/appsAdminPage';
-import { EditGroupAdminPage } from './../../pageObjectModels/editGroupAdminPage';
-import { EditRoleAdminPage } from './../../pageObjectModels/editRoleAdminPage';
-import { EditUserAdminPage } from './../../pageObjectModels/editUserAdminPage';
-import { GroupsSecurityAdminPage } from './../../pageObjectModels/groupsSecurityAdminPage';
-import { RolesSecurityAdminPage } from './../../pageObjectModels/rolesSecurityAdminPage';
-import { UsersSecurityAdminPage } from './../../pageObjectModels/usersSecurityAdminPage';
 
 type AppTestFixtures = {
   adminHomePage: AdminHomePage;
